csv tests

+++ x-pack/plugin/esql/src/internalClusterTest/java/org/elasticsearch/xpack/esql/plugin/QueryStringFunctionIT.java
@@ -170,11 +170,11 @@ public class QueryStringFunctionIT extends AbstractEsqlIntegTestCase {
             assertThat(resp.columns().stream().map(ColumnInfoImpl::name).toList(), equalTo(List.of("id", "_score")));
             assertThat(
                 resp.columns().stream().map(ColumnInfoImpl::type).map(DataType::toString).toList(),
-                equalTo(List.of("INTEGER", "FLOAT"))
+                equalTo(List.of("INTEGER", "DOUBLE"))
             );
             // values
             List<List<Object>> values = getValuesList(resp);
-            assertMap(values, matchesList().item(List.of(1, 1.1565589F)).item(List.of(6, 0.9114002F)));
+            assertMap(values, matchesList().item(List.of(1, 1.1565589D)).item(List.of(6, 0.9114002D)));
         }
     }
 
@@ -192,11 +192,11 @@ public class QueryStringFunctionIT extends AbstractEsqlIntegTestCase {
             assertThat(resp.columns().stream().map(ColumnInfoImpl::name).toList(), equalTo(List.of("id", "_score")));
             assertThat(
                 resp.columns().stream().map(ColumnInfoImpl::type).map(DataType::toString).toList(),
-                equalTo(List.of("INTEGER", "FLOAT"))
+                equalTo(List.of("INTEGER", "DOUBLE"))
             );
             // values
             List<List<Object>> values = getValuesList(resp);
-            assertMap(values, matchesList().item(List.of(1, 1.1565589F)).item(List.of(6, 0.9114002F)));
+            assertMap(values, matchesList().item(List.of(1, 1.1565589D)).item(List.of(6, 0.9114002D)));
         }
     }
 }
